Details of Frog Puzzle Game
The objective of the puzzle is to switch the positions of all the colored frogs, so that all three green frogs are on the right and all three brown frogs are on the left. When all the frogs have reached the other side of the pond, you win!

Enter the position of a frog to make it jump forward. A frog can jump a maximum of two spaces, and the destination spot must be free.

Green frogs can only hop to the right. Brown frogs can only hop to the left. Frogs can only jump onto an empty rock("-"). A frog can jump over one frog to an empty rock("-")—but only over one.

Frogs cannot jump back, only forward. 

If you get stuck and your frogs can't jump anywhere, reset the game.

You can play an animated version of this game here:
https://data.bangtech.com/algorithm/switch_frogs_to_the_opposite_side.htm